refactored cluster page header
